+++ compiler-cli/src/new/tests.rs
@@ -11,7 +11,7 @@ fn new() {
     let creator = super::Creator::new(
         super::NewOptions {
             project_root: path.to_string(),
-            template: super::Template::Lib,
+            template: super::Template::Erlang,
             name: None,
             skip_git: false,
             skip_github: false,
@@ -34,6 +34,62 @@ fn new() {
 }
 
 #[test]
+fn new_with_default_template() {
+    let tmp = tempfile::tempdir().unwrap();
+    let path = Utf8PathBuf::from_path_buf(tmp.into_path()).expect("Non Utf8 Path");
+
+    let creator = super::Creator::new(
+        super::NewOptions {
+            project_root: path.join("my_project").to_string(),
+            template: super::Template::Erlang,
+            name: None,
+            skip_git: false,
+            skip_github: true,
+        },
+        "1.0.0-gleam",
+    )
+    .unwrap();
+    creator.run().unwrap();
+
+    insta::glob!(path, "my_project/*.*", |file_path| {
+        if !file_path.is_dir() {
+            insta::assert_snapshot!(crate::fs::read(
+                Utf8PathBuf::from_path_buf(file_path.to_path_buf()).expect("Non Utf8 Path"),
+            )
+            .unwrap());
+        }
+    });
+}
+
+#[test]
+fn new_with_javascript_template() {
+    let tmp = tempfile::tempdir().unwrap();
+    let path = Utf8PathBuf::from_path_buf(tmp.into_path()).expect("Non Utf8 Path");
+
+    let creator = super::Creator::new(
+        super::NewOptions {
+            project_root: path.join("my_project").to_string(),
+            template: super::Template::JavaScript,
+            name: None,
+            skip_git: false,
+            skip_github: true,
+        },
+        "1.0.0-gleam",
+    )
+    .unwrap();
+    creator.run().unwrap();
+
+    insta::glob!(path, "my_project/*.*", |file_path| {
+        if !file_path.is_dir() {
+            insta::assert_snapshot!(crate::fs::read(
+                Utf8PathBuf::from_path_buf(file_path.to_path_buf()).expect("Non Utf8 Path"),
+            )
+            .unwrap());
+        }
+    });
+}
+
+#[test]
 fn new_with_skip_git() {
     let tmp = tempfile::tempdir().unwrap();
     let path = Utf8PathBuf::from_path_buf(tmp.path().join("my_project")).expect("Non Utf8 Path");
@@ -41,7 +97,7 @@ fn new_with_skip_git() {
     let creator = super::Creator::new(
         super::NewOptions {
             project_root: path.to_string(),
-            template: super::Template::Lib,
+            template: super::Template::Erlang,
             name: None,
             skip_git: true,
             skip_github: false,
@@ -62,7 +118,7 @@ fn new_with_skip_github() {
     let creator = super::Creator::new(
         super::NewOptions {
             project_root: path.to_string(),
-            template: super::Template::Lib,
+            template: super::Template::Erlang,
             name: None,
             skip_git: false,
             skip_github: true,
@@ -86,7 +142,7 @@ fn new_with_skip_git_and_github() {
     let creator = super::Creator::new(
         super::NewOptions {
             project_root: path.to_string(),
-            template: super::Template::Lib,
+            template: super::Template::Erlang,
             name: None,
             skip_git: true,
             skip_github: true,
@@ -110,7 +166,7 @@ fn invalid_path() {
     assert!(super::Creator::new(
         super::NewOptions {
             project_root: path.to_string(),
-            template: super::Template::Lib,
+            template: super::Template::Erlang,
             name: None,
             skip_git: false,
             skip_github: false,
@@ -128,7 +184,7 @@ fn invalid_name() {
     assert!(super::Creator::new(
         super::NewOptions {
             project_root: path.to_string(),
-            template: super::Template::Lib,
+            template: super::Template::Erlang,
             name: Some("-".into()),
             skip_git: false,
             skip_github: false,
@@ -148,7 +204,7 @@ fn existing_directory_no_files() {
     let creator = super::Creator::new(
         super::NewOptions {
             project_root: path.to_string(),
-            template: super::Template::Lib,
+            template: super::Template::Erlang,
             name: None,
             skip_git: true,
             skip_github: true,
@@ -175,7 +231,7 @@ fn existing_directory_with_one_existing_file() {
     assert!(super::Creator::new(
         super::NewOptions {
             project_root: path.to_string(),
-            template: super::Template::Lib,
+            template: super::Template::Erlang,
             name: None,
             skip_git: true,
             skip_github: true,
@@ -198,7 +254,7 @@ fn existing_directory_with_non_generated_file() {
     let creator = super::Creator::new(
         super::NewOptions {
             project_root: path.to_string(),
-            template: super::Template::Lib,
+            template: super::Template::Erlang,
             name: None,
             skip_git: true,
             skip_github: true,
@@ -228,7 +284,7 @@ fn conflict_with_existing_files() {
         super::Creator::new(
             super::NewOptions {
                 project_root: path.to_string(),
-                template: super::Template::Lib,
+                template: super::Template::Erlang,
                 name: None,
                 skip_git: true,
                 skip_github: true,